assignment
In this assignment, you will create a flashcard app for teaching addition and multiplication. The student will choose the type of problem, how many they want to solve, and the maximum operand to include in each problem. The program will keep track of how many problems the student answered correctly and report that number along with their percentage. It will then repeat the process, allowing the student to choose a different kind of problem for the next round.

Here is an example of how the program should work:
Choose the type of problem: 
1. Addition
2. Multiplication
3. Mixed
4. Quit
Enter your choice: 3
How many problems? 5
Largest operand? 10

1 * 1 = ? 2
Incorrect. The answer is 1.
1 + 7 = ? 8
Correct!
6 + 0 = ? 6
Correct!
4 + 9 = ? 12
Incorrect. The answer is 13.
4 * 0 = ? 0
Correct!

You answered 3 out of 5 correctly.
Your average was 60.00.